Game is a dataclass
Make Game a dataclass to help me catch any mismatch between attributes
and what is defined in '__init__()'. As a dataclass, '__init__()' is
generated for me, so I only need to manually set the attributes that
depend on other attributes. Set these attributes to 'field(init=False)'
so that the caller ('main.py') is not responsible to pass these in. Set
the values in '__post_init__()'.

Also renamed 'engine/geometry_operators.py' to 'engine/coord_xfm.py' in
keeping with the pattern that if the file defines a single class, then
the file name matches the class name.