Add class Ticks, pause, and do a minor refactor
Ticks is a container for all the TickCounters. These are counters that
increment after some number of video frames. Ticks maintains a count of
the number of video frames. The individual TickCounters in Ticks are
responsible for updating themselves when their period elapses (when the
number of frames modulo the counter's period equals zero).

Pressing space toggles pausing the frame counter. Pausing the frame
counter causes all TickCounters to pause as well.

Minor refactoring:

- Make sure all attributes added in __post_init__() also show up in the
  list of attributes at the top of the dataclass.
- Eliminate redundancy between __str__() and fmt() methods by having
  __str__() call fmt(). Only use a default precision with __str__(): if
  the caller uses fmt(), then a precision is required.