Major refactor -- clean up debug HUD snapshots
I had a complicated system before for taking snapshots of values. It
limited me to looking at only one section of code. I realized the
snapshots could simply be a dict of variable-name keys and print-message
values. I simply print whatever is in the dict.

I also realized I was coupling the idea of reading FPS
infrequently to the idea of debug HUD prints that do not happen on every
iteration. I decoupled these ideas and plan to address rapidly changing
values (like FPS) by creating a data buffer that I can write/read as
often as I want, but it internally tracks when it is time to update the
value output when I read from it (perhaps based on how often I write to
it, or maybe it just stores the N most-recent values and averages them).

I cleaned up the debug HUD print formatting to make it much easier to
read what is happening. And I changed the font from Roboto to Proggy,
which is in a 'fonts/' folder that is now part of the symbolic links.
Now that I am using a local 'fonts/', the debug HUD font is selected in
'game.py' by setting the attribute debug_font to the local 'fonts/'
path.