Split Timing ticks into "video" and "game"
Pausing the game also paused the frame count. This meant the HUD FPS
stopped updating. So instead of one instance of Ticks in Timing, I made
two instances: video_ticks and game_ticks. And now the Ticks constructor
takes a clock (str) which is either "video" or "game". Ticks uses this
is the __post_init__() in a match-case to see which dictionary
(video_counter or game_counter) to assign to this instance.